requirements.txt
tensorflow>=2.12.0
streamlit>=1.28.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
